+++ Engine/shader/texVertex.vert
@@ -6,7 +6,7 @@ precision mediump float;
 
 layout(location = 0) in vec4 i_pos;
 layout(location = 1) in vec4 i_col;
-layout(location = 2) in float index;
+layout(location = 2) in int index;
 layout(location = 3) in float isTex;
 
 layout(location = 0) out vec2 o_uv;
@@ -32,9 +32,9 @@ void main()
     o_uv.x = ((i_pos.x + 1) / 2);
     o_uv.y = ((i_pos.y + 1) / 2);
 
-    outTexIndex = matrix[int(index)].texIndex;
+    outTexIndex = matrix[index].texIndex;
     outIsTex = isTex;
     o_col = i_col;
 
-    gl_Position = matrix[int(index)].projection *  matrix[int(index)].view * matrix[int(index)].model * i_pos;
+    gl_Position = matrix[index].projection *  matrix[index].view * matrix[index].model * i_pos;
 }
